USB descriptors in its own file
Moved USB HID descriptors and related constants to new descriptors.c/h file
