+++ usb/descriptors.c
@@ -1,0 +1,77 @@
+#include "descriptors.h"
+
+/* Device descriptor */
+const struct usb_device_descriptor hid_device_desc = {
+    .bLength = sizeof(struct usb_device_descriptor),
+    .bDescriptorType = USB_DTYPE_DEVICE,
+    .bcdUSB = VERSION_BCD(2, 0, 0),
+    .bDeviceClass = USB_CLASS_PER_INTERFACE,
+    .bDeviceSubClass = USB_SUBCLASS_NONE,
+    .bDeviceProtocol = USB_PROTO_NONE,
+    .bMaxPacketSize0 = USB_EP0_SIZE,
+    .idVendor = HID_VID_TOYPAD,
+    .idProduct = HID_PID_TOYPAD,
+    .bcdDevice = VERSION_BCD(1, 0, 0),
+    .iManufacturer = 1,
+    .iProduct = 2,
+    .iSerialNumber = 3,
+    .bNumConfigurations = 1,
+};
+
+/* Device configuration descriptor */
+const struct HidConfigDescriptor hid_cfg_desc = {
+    .config =
+        {
+            .bLength = sizeof(struct usb_config_descriptor),
+            .bDescriptorType = USB_DTYPE_CONFIGURATION,
+            .wTotalLength = sizeof(struct HidConfigDescriptor),
+            .bNumInterfaces = 1,
+            .bConfigurationValue = 1,
+            .iConfiguration = NO_DESCRIPTOR,
+            .bmAttributes = USB_CFG_ATTR_RESERVED,
+            .bMaxPower = USB_CFG_POWER_MA(500),
+        },
+    .intf_0 =
+        {
+            .hid =
+                {
+                    .bLength = sizeof(struct usb_interface_descriptor),
+                    .bDescriptorType = USB_DTYPE_INTERFACE,
+                    .bInterfaceNumber = 0,
+                    .bAlternateSetting = 0,
+                    .bNumEndpoints = 2,
+                    .bInterfaceClass = USB_CLASS_HID,
+                    .bInterfaceSubClass = USB_HID_SUBCLASS_NONBOOT,
+                    .bInterfaceProtocol = USB_HID_PROTO_NONBOOT,
+                    .iInterface = NO_DESCRIPTOR,
+                },
+            .hid_desc =
+                {
+                    .bLength = sizeof(struct usb_hid_descriptor),
+                    .bDescriptorType = USB_DTYPE_HID,
+                    .bcdHID = VERSION_BCD(1, 0, 0),
+                    .bCountryCode = USB_HID_COUNTRY_NONE,
+                    .bNumDescriptors = 1,
+                    .bDescriptorType0 = USB_DTYPE_HID_REPORT,
+                    .wDescriptorLength0 = sizeof(hid_report_desc),
+                },
+            .hid_ep_in =
+                {
+                    .bLength = sizeof(struct usb_endpoint_descriptor),
+                    .bDescriptorType = USB_DTYPE_ENDPOINT,
+                    .bEndpointAddress = HID_EP_IN,
+                    .bmAttributes = USB_EPTYPE_INTERRUPT,
+                    .wMaxPacketSize = HID_EP_SZ,
+                    .bInterval = HID_INTERVAL,
+                },
+            .hid_ep_out =
+                {
+                    .bLength = sizeof(struct usb_endpoint_descriptor),
+                    .bDescriptorType = USB_DTYPE_ENDPOINT,
+                    .bEndpointAddress = HID_EP_OUT,
+                    .bmAttributes = USB_EPTYPE_INTERRUPT,
+                    .wMaxPacketSize = HID_EP_SZ,
+                    .bInterval = HID_INTERVAL,
+                },
+        },
+};
