+++ usb/descriptors.h
@@ -1,0 +1,68 @@
+#pragma once
+#include <furi.h>
+#include <usb.h>
+#include <usb_hid.h>
+
+#define HID_VID_TOYPAD 0x0e6f // Logic3
+#define HID_PID_TOYPAD 0x0241
+
+#define HID_EP_IN  0x81
+#define HID_EP_OUT 0x01
+
+#define HID_EP_SZ 0x20 // 32 bytes packet size
+
+#define HID_INTERVAL 1
+
+#define USB_EP0_SIZE 64
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+/* String descriptors */
+enum UsbDevDescStr {
+    UsbDevLang = 0,
+    UsbDevManuf = 1,
+    UsbDevProduct = 2,
+    UsbDevSerial = 3,
+};
+
+struct HidIntfDescriptor {
+    struct usb_interface_descriptor hid;
+    struct usb_hid_descriptor hid_desc;
+    struct usb_endpoint_descriptor hid_ep_in;
+    struct usb_endpoint_descriptor hid_ep_out;
+};
+
+struct HidConfigDescriptor {
+    struct usb_config_descriptor config;
+    struct HidIntfDescriptor intf_0;
+} __attribute__((packed));
+
+/* HID report descriptor */
+static const uint8_t hid_report_desc[] = {
+    0x06, 0x00, 0xFF, // Usage Page (Vendor Defined)
+    0x09, 0x01, // Usage (Vendor Usage 1)
+    0xA1, 0x01, // Collection (Application)
+    0x19, 0x01, //   Usage Minimum (Vendor Usage 1)
+    0x29, 0x20, //   Usage Maximum (Vendor Usage 32)
+    0x15, 0x00, //   Logical Minimum (0)
+    0x26, 0xFF, 0x00, //   Logical Maximum (255)
+    0x75, 0x08, //   Report Size (8 bits)
+    0x95, 0x20, //   Report Count (32 bytes)
+    0x81, 0x00, //   Input (Data, Array, Absolute)
+    0x19, 0x01, //   Usage Minimum (Vendor Usage 1)
+    0x29, 0x20, //   Usage Maximum (Vendor Usage 32)
+    0x91, 0x00, //   Output (Data, Array, Absolute)
+    0xC0 // End Collection
+};
+
+/* Device descriptor */
+extern const struct usb_device_descriptor hid_device_desc;
+
+/* Device configuration descriptor */
+extern const struct HidConfigDescriptor hid_cfg_desc;
+
+#ifdef __cplusplus
+}
+#endif
